:art: :: button components 위치 변경
